task: implement Saleable V.4
Human cash and price paramether in sell() is now double.
Fixed misplaced message of succesful animal buyment.
Buyment of device now result in money transaction, nut still lacks of any functionality.

+++ src/com/company/Human.java
@@ -10,7 +10,7 @@ public class Human extends  Animal{
     public Animal pet;
     private Double salary = 1500.0d;
     private Car setOfWheels;
-    public float cash;
+    public double cash;
 
     public Human() {
         super("Homo sapient");
